fix: 修复解包 GDK 时文件平铺的 bug

+++ BedrockBoot.Linux/Entry/Progress/DownloadProgress.cs
@@ -9,6 +9,6 @@ public class DownloadProgress
         ? Math.Min(100, Math.Round((double)DownloadedBytes / TotalBytes * 100, 2))
         : 0;
 
-    public double BytesPerSecond { get; set; }
+    public long BytesPerSecond { get; set; }
     public double EstimatedRemainingSeconds { get; set; }
 }
